refactor(earn): OfferCard to own file
