+++ src/components/earn/EarnPage.tsx
@@ -1,19 +1,9 @@
-import { useId, useMemo, useState, type PropsWithChildren } from 'react'
+import { useId, useMemo, useState } from 'react'
 import { yupResolver } from '@hookform/resolvers/yup'
 import { startmakerMutation, stopmakerOptions } from '@joinmarket-webui/joinmarket-api-ts/@tanstack/react-query'
-import type { ErrorMessage, SessionResponse, StartMakerRequest } from '@joinmarket-webui/joinmarket-api-ts/jm'
+import type { ErrorMessage, StartMakerRequest } from '@joinmarket-webui/joinmarket-api-ts/jm'
 import { useMutation, useQuery } from '@tanstack/react-query'
-import type { TFunction } from 'i18next'
-import {
-  AlertTriangle,
-  FingerprintIcon,
-  HandCoinsIcon,
-  HandshakeIcon,
-  Loader2Icon,
-  Maximize2Icon,
-  Minimize2Icon,
-  PercentIcon,
-} from 'lucide-react'
+import { AlertTriangle, HandshakeIcon, Loader2Icon, PercentIcon } from 'lucide-react'
 import { useForm, useWatch } from 'react-hook-form'
 import type { Resolver, SubmitHandler } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
@@ -30,18 +20,15 @@ import { Tabs, TabsContent } from '@/components/ui/tabs'
 import { OFFER_FEE_ABS_MIN, OFFER_FEE_REL_MIN, OFFER_MINSIZE_MIN } from '@/constants/jam'
 import type { OfferType } from '@/constants/jm'
 import { useApiClient } from '@/hooks/useApiClient'
-import { useDisplaySettings } from '@/hooks/useDisplaySettings'
 import { useFeeConfigValidation } from '@/hooks/useFeeConfigValidation'
 import { useRefreshSession } from '@/hooks/useRefreshSession'
 import { withQueryDelay } from '@/lib/queryClient'
 import { cn, factorToPercentage, isAbsoluteOffer, isRelativeOffer, percentageToFactor } from '@/lib/utils'
 import type { WalletFileName } from '@/lib/utils'
 import { jmSessionStore } from '@/store/jmSessionStore'
-import type { Milliseconds, Unbox } from '@/types/global'
-import { CurrencySymbol, SatSymbol } from '../CurrencySymbol'
-import { Badge } from '../ui/badge'
-import { Card, CardContent, CardAction, CardDescription, CardFooter, CardHeader, CardTitle } from '../ui/card'
-import { Tooltip, TooltipContent, TooltipTrigger } from '../ui/tooltip'
+import type { Milliseconds } from '@/types/global'
+import { SatSymbol } from '../CurrencySymbol'
+import { OfferCard } from './OfferCard'
 
 // In order to prevent state mismatch, the 'maker stop' response is delayed shortly.
 // Even though the API response suggests that the maker has started or stopped immediately, it seems that this is not always the case.
@@ -443,7 +430,7 @@ export const EarnPage = ({ walletFileName }: EarnPageProps) => {
       {jmSessionState.offer_list && jmSessionState.offer_list.length > 0 && (
         <>
           <div className="space-y-2">
-            <CurrentOffer value={jmSessionState.offer_list[0]} nickname={jmSessionState.nickname}>
+            <OfferCard value={jmSessionState.offer_list[0]} nickname={jmSessionState.nickname}>
               <Button type="button" onClick={() => onStop()} className="w-full" size="lg">
                 {isWaitingMakerStop ? (
                   <>
@@ -454,7 +441,7 @@ export const EarnPage = ({ walletFileName }: EarnPageProps) => {
                   <>{t('earn.button_stop')}</>
                 )}
               </Button>
-            </CurrentOffer>
+            </OfferCard>
           </div>
         </>
       )}
@@ -482,113 +469,5 @@ export const EarnPage = ({ walletFileName }: EarnPageProps) => {
         walletFileName={walletFileName}
       />
     </div>
-  )
-}
-
-const OfferTypeBadge = ({ value }: { value: Unbox<SessionResponse['offer_list']> }) => {
-  const { t } = useTranslation()
-  const text = renderOfferText(value, t)
-  return <Badge variant={text ? 'default' : 'outline'}>{text}</Badge>
-}
-
-const renderOfferText = (value: Unbox<SessionResponse['offer_list']>, t: TFunction<'translation', undefined>) => {
-  if (isAbsoluteOffer(value?.ordertype || '')) {
-    return t('earn.current.text_offer_type_absolute')
-  }
-  if (isRelativeOffer(value?.ordertype || '')) {
-    return t('earn.current.text_offer_type_relative')
-  }
-  return value?.ordertype
-}
-
-interface CurrentOfferProps {
-  value: Unbox<SessionResponse['offer_list']>
-  nickname: SessionResponse['nickname']
-}
-
-function CurrentOffer({ value, nickname, children }: PropsWithChildren<CurrentOfferProps>) {
-  const { t } = useTranslation()
-  const { currency, formatAmount } = useDisplaySettings()
-
-  return (
-    <Card className="w-full">
-      <CardHeader>
-        <CardTitle>{t('earn.current.text_offer')}</CardTitle>
-        <CardDescription></CardDescription>
-        <CardAction>
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <OfferTypeBadge value={value} />
-            </TooltipTrigger>
-            <TooltipContent>{value?.ordertype}</TooltipContent>
-          </Tooltip>
-        </CardAction>
-      </CardHeader>
-      <CardContent className="grid gap-4 sm:grid-cols-2">
-        <div className="flex items-center gap-4">
-          <FingerprintIcon />
-          <div className="flex flex-col">
-            <span className="text-muted-foreground text-sm font-semibold">
-              {
-                /*TODO: i18n*/
-                t('Offer Id')
-              }
-            </span>
-            <span className="text-md font-mono">
-              {nickname}:{value?.oid}
-            </span>
-          </div>
-        </div>
-
-        <div className="flex items-center gap-4">
-          <HandCoinsIcon />
-          <div className="flex flex-col">
-            <span className="text-muted-foreground text-sm font-semibold">{t('earn.current.text_cjfee')}</span>
-            <span className="text-sm">
-              {isRelativeOffer(value?.ordertype || '') ? (
-                <>{factorToPercentage(parseFloat(value?.cjfee || '') || 0)}%</>
-              ) : (
-                <>
-                  <span className="tabular-nums">{formatAmount(parseInt(String(value?.cjfee || '0'), 10))}</span>
-                  <CurrencySymbol currency={currency} isPrivate={false} size="sm" />
-                </>
-              )}
-            </span>
-          </div>
-        </div>
-        <div className="flex items-center gap-4">
-          <Minimize2Icon />
-          <div className="flex flex-col">
-            <span className="text-muted-foreground text-sm font-semibold">{t('earn.current.text_minsize')}</span>
-            <span className="text-sm">
-              <span className="tabular-nums">{formatAmount(parseInt(String(value?.minsize || '0'), 10))}</span>
-              <CurrencySymbol currency={currency} isPrivate={false} size="sm" />
-            </span>
-          </div>
-        </div>
-        <div className="flex items-center gap-4">
-          <Maximize2Icon />
-          <div className="flex flex-col">
-            <span className="text-muted-foreground text-sm font-semibold">{t('earn.current.text_maxsize')}</span>
-            <span className="text-sm">
-              <span className="tabular-nums">{formatAmount(parseInt(String(value?.maxsize || '0'), 10))}</span>
-              <CurrencySymbol currency={currency} isPrivate={false} size="sm" />
-            </span>
-          </div>
-        </div>
-        {!!value?.txfee && (
-          <div className="flex items-center gap-4">
-            <div className="flex flex-col">
-              <span className="text-muted-foreground text-sm font-semibold">{t('earn.current.text_txfee')}</span>
-              <span className="text-muted-foreground text-sm">
-                <span className="tabular-nums">{formatAmount(parseInt(String(value?.txfee || '0'), 10))}</span>
-                <CurrencySymbol currency={currency} isPrivate={false} size="sm" />
-              </span>
-            </div>
-          </div>
-        )}
-      </CardContent>
-      <CardFooter className="flex-col gap-2">{children}</CardFooter>
-    </Card>
   )
 }
